Move everything to a `gamerules` package
Move game rule entry factory into its own class
Rename fields to better match their usage
